here are some changes ai agents suggested
