+++ src/components/LabelModel/LabelModel.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState, type FC } from "react";
+import { useEffect, useState, type FC, Suspense } from "react";
 import { getCalendarCell, labelEvent } from "../../store/calendar";
 import { FormControl } from "../FormControl/FormControl";
 import { Modal } from "../Modal";
@@ -26,46 +26,48 @@ export const LabelModel: FC<Props> = ({ id, onClose }) => {
   };
 
   return (
-    <Modal
-      id="label"
-      footer={
-        <div className="flex w-full flex-col-reverse justify-end gap-2 sm:flex-row">
-          <button
-            id="modal-cancel-btn"
-            type="button"
-            className="btn btn-outline w-full sm:w-auto"
-            onClick={onClose}
-          >
-            Cancel
-          </button>
-          <button className="btn btn-primary" onClick={onSave}>
-            Set Label
-          </button>
-        </div>
-      }
-      title="Cell Label"
-      isOpen={isOpen}
-      onClose={onClose}
-    >
-      <FormControl id="label-input" label="Cell Label">
-        <input
-          id="label-input"
-          onChange={(e) => {
-            setValue(e.target.value);
-          }}
-          onKeyDown={(e) => {
-            if (e.key === "Enter") {
-              onSave();
-            }
-          }}
-          value={value}
-          required
-          aria-required="true"
-          type="text"
-          placeholder="Label"
-          className="input w-full"
-        />
-      </FormControl>
-    </Modal>
+    <Suspense fallback={null}>
+      <Modal
+        id="label"
+        footer={
+          <div className="flex w-full flex-col-reverse justify-end gap-2 sm:flex-row">
+            <button
+              id="modal-cancel-btn"
+              type="button"
+              className="btn btn-outline w-full sm:w-auto"
+              onClick={onClose}
+            >
+              Cancel
+            </button>
+            <button className="btn btn-primary" onClick={onSave}>
+              Set Label
+            </button>
+          </div>
+        }
+        title="Cell Label"
+        isOpen={isOpen}
+        onClose={onClose}
+      >
+        <FormControl id="label-input" label="Cell Label">
+          <input
+            id="label-input"
+            onChange={(e) => {
+              setValue(e.target.value);
+            }}
+            onKeyDown={(e) => {
+              if (e.key === "Enter") {
+                onSave();
+              }
+            }}
+            value={value}
+            required
+            aria-required="true"
+            type="text"
+            placeholder="Label"
+            className="input w-full"
+          />
+        </FormControl>
+      </Modal>
+    </Suspense>
   );
 };
